refactor(settings): move security components to features
